requirements.txt
# Global dependencies of the project
fire
fastapi
uvicorn
pyyaml

# demo dependencies
